+++ fields/fields.py
@@ -1,4 +1,5 @@
 import pygame
+import universal.fonts as fonts
 
-def field_placement():
-    pass
+def field_placement(screen, x, y):
+    pygame.draw.rect(screen, "white", pygame.Rect(x, y, 125, 125))