+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - A minor issue has been corrected within that dev env setup guide.
 - The dev env setup guide is fairly finished (for) now.
 - The first text is being displayed.
+- The field function is sort of there.